refactor(addons): remove validate.yaml and use values.schema.json
This commit refactors the validation mechanism for addon configurations by removing the custom  files.

Validation logic has been migrated to the standard  for the following addons:
- clickhouse
- oceanbase-ce
- pulsar

This change centralizes validation within the Helm standard, improving maintainability and clarity. It also includes a minor typo fix in the Elasticsearch validation template.